Move all utils out of main component folder
