+++ editor.planx.uk/src/ui/editor/RichTextInput/RichTextInput.tsx
@@ -8,241 +8,37 @@ import FormatListNumbered from "@mui/icons-material/FormatListNumbered";
 import LinkIcon from "@mui/icons-material/Link";
 import Box from "@mui/material/Box";
 import IconButton from "@mui/material/IconButton";
-import { type Editor, type JSONContent } from "@tiptap/core";
-import Bold from "@tiptap/extension-bold";
-import BulletList from "@tiptap/extension-bullet-list";
-import Document from "@tiptap/extension-document";
-import HardBreak from "@tiptap/extension-hard-break";
-import Heading from "@tiptap/extension-heading";
+import { type Editor } from "@tiptap/core";
 import History from "@tiptap/extension-history";
-import Italic from "@tiptap/extension-italic";
-import Link from "@tiptap/extension-link";
-import ListItem from "@tiptap/extension-list-item";
 import Mention from "@tiptap/extension-mention";
-import OrderedList from "@tiptap/extension-ordered-list";
-import Paragraph from "@tiptap/extension-paragraph";
 import ExtensionPlaceholder from "@tiptap/extension-placeholder";
-import Text from "@tiptap/extension-text";
-import { generateHTML, generateJSON } from "@tiptap/html";
-import { EditorContent, ReactRenderer, useEditor } from "@tiptap/react";
-import { map } from "ramda";
+import { EditorContent, useEditor } from "@tiptap/react";
 import React, {
   ChangeEvent,
   type FC,
-  forwardRef,
   useCallback,
   useEffect,
-  useImperativeHandle,
   useRef,
   useState,
 } from "react";
-import tippy, { type Instance } from "tippy.js";
-import { create } from "zustand";
 
 import Input from "../../shared/Input/Input";
 import PublicFileUploadButton from "../../shared/PublicFileUploadButton";
-import CustomImage from "../RichTextImage";
 import { PopupError } from "./components/PopUpError";
+import { suggestion } from "./components/suggestion";
+import { RichContentContainer, StyledBubbleMenu } from "./styles";
+import { commonExtensions } from "./tiptapExtensions";
+import { Props } from "./types";
 import {
-  MentionItems,
-  MentionItemsButton,
-  MentionItemsEmpty,
-  RichContentContainer,
-  StyledBubbleMenu,
-} from "./styles";
-import { MentionListProps, Props } from "./types";
-
-const passportClassName = "passport";
-
-// Shared tiptap editor extensions
-const commonExtensions = [
-  Document,
-  Paragraph,
-  Text,
-  Bold,
-  Italic,
-  HardBreak,
-  Link.configure({
-    openOnClick: false,
-    autolink: false,
-  }),
-  Heading.configure({
-    levels: [1, 2, 3],
-  }),
-  BulletList,
-  OrderedList,
-  ListItem,
-  CustomImage,
-];
-
-// Tiptap editor extensions used to convert between HTML and Prosemirror document state (used internally by tiptap)
-const conversionExtensions = [
-  ...commonExtensions,
-  Mention.configure({
-    HTMLAttributes: {
-      class: passportClassName,
-    },
-  }),
-];
-
-interface VariablesState {
-  variables: string[];
-  addVariable: (newVariable: string) => void;
-}
-
-export const emptyContent = "<p></p>";
-
-// Specify whether a selection is unsuitable for ensuring accessible links
-const linkSelectionError = (selectionHtml: string): string | null => {
-  if (selectionHtml.startsWith("<p>") && selectionHtml.endsWith("</p>")) {
-    const text = selectionHtml.slice(3, -4);
-    const lowercaseText = text.toLowerCase().trim().replace(/[.,]/g, "");
-    if (lowercaseText === "click here" || lowercaseText === "clicking here") {
-      return "Links must be set over text that accurately describes what the link is for. Avoid generic language such as 'click here'.";
-    }
-    if (text[0] && text[0] !== text[0].toUpperCase() && text.length < 8) {
-      return "Make sure the link text accurately describes the what the link is for.";
-    }
-  }
-  return null;
-};
-
-// Maintain a store of variables as they are created in the '@'-mention plugin, making them available in memory for next time.
-// TODO: explore instantiating from a hard-coded list, or persisting in either the backend or local storage.
-const useVariablesStore = create<VariablesState>()((set) => ({
-  variables: [],
-  addVariable: (newVariable: string) =>
-    set((state) => ({ variables: [...state.variables, newVariable] })),
-}));
-
-export const toHtml = (doc: JSONContent) => {
-  const outgoingHtml = generateHTML(doc, conversionExtensions);
-  return outgoingHtml === emptyContent ? "" : outgoingHtml;
-};
-
-export const fromHtml = (htmlString: string) => {
-  return generateJSON(
-    htmlString === "" ? emptyContent : htmlString,
-    conversionExtensions,
-  );
-};
-
-export const injectVariables = (
-  htmlString: string,
-  vars: Record<string, string>,
-) => {
-  const doc = fromHtml(htmlString);
-  return toHtml(
-    modifyDeep((node) => {
-      return node.type === "mention"
-        ? {
-            ...node,
-            type: "text",
-            text: vars[node.attrs.id] || "Unknown",
-            attrs: undefined,
-          }
-        : null;
-    })(doc),
-  );
-};
-
-/**
- * Traverse a nested object/array and apply a modification at each level. If the modifier returns `null`, it leaves the result unchanged.
- * Used to inject placeholder values into a document structure.
- */
-export const modifyDeep =
-  (fn: (field: Value) => Value) =>
-  (val: Value): Value => {
-    if (!val) {
-      return val;
-    }
-    const mod = fn(val);
-    if (mod) {
-      return mod;
-    }
-    if (Array.isArray(val)) {
-      return map(modifyDeep(fn), val);
-    }
-    if (typeof val === "object") {
-      return map(modifyDeep(fn), val);
-    }
-    return val;
-  };
-
-// Generic value that `modifyDeep` works off of. Since it can be object, array or primitive, any typing more accurate than `any` is not compiling at the moment.
-// TODO: find a better typing alternative
-type Value = any;
-
-const initialUrlValue = "https://";
-
-// Makes sure that if the user pastes a full URL into the input, the pre-populated `https://` is removed
-// e.g. https://https://something.com -> https://something.com
-// Also trim whitespace from links which flag sanitation errors
-const trimUrlValue = (url: string) => {
-  if (url.startsWith(`${initialUrlValue}${initialUrlValue}`)) {
-    return url.slice(initialUrlValue.length);
-  }
-  return url.trim();
-};
-
-const getContentHierarchyError = (doc: JSONContent): string | null => {
-  let h1Index = -1;
-  let h2Index = -1;
-
-  let error: string | null = null;
-
-  (doc.content || []).forEach((d: JSONContent, index) => {
-    if (d.type === "paragraph") {
-      return;
-    } else if (d.type === "heading") {
-      const level = d.attrs?.level === 1 ? 1 : 2;
-      if (level === 1) {
-        if (h1Index === -1 && h2Index !== -1) {
-          error = "A level 1 heading must come before a level 2 heading.";
-        } else if (h1Index !== -1) {
-          error =
-            "There cannot be more than one level 1 heading in the document.";
-        } else if (index !== 0) {
-          error = "The level 1 heading must come first in the document.";
-        }
-        h1Index = index;
-        return;
-      }
-      if (level === 2) {
-        if (h1Index === -1) {
-          error = "A level 1 heading must come before a level 2 heading.";
-        }
-        h2Index = index;
-        return;
-      }
-    }
-    return null;
-  });
-
-  return error;
-};
-
-const getLinkNewTabError = (
-  content: JSONContent | undefined = [],
-): string | undefined => {
-  let error: string | undefined;
-  if (!content) return;
-
-  content.forEach((child: JSONContent) => {
-    if (!child.content) return;
-
-    child.content.forEach(({ marks, text }) => {
-      const isLink = marks?.map(({ type }) => type).includes("link");
-      const hasOpenTabText = text?.includes("(opens in a new tab)");
-
-      if (hasOpenTabText && !isLink) {
-        error = 'Links must wrap the text "(opens in a new tab)"';
-      }
-    });
-  });
-
-  return error;
-};
+  fromHtml,
+  getContentHierarchyError,
+  getLinkNewTabError,
+  initialUrlValue,
+  linkSelectionError,
+  passportClassName,
+  toHtml,
+  trimUrlValue,
+} from "./utils";
 
 const RichTextInput: FC<Props> = (props) => {
   const stringValue = String(props.value || "");
@@ -569,154 +365,5 @@ const RichTextInput: FC<Props> = (props) => {
     </RichContentContainer>
   );
 };
-
-// Implemented based on the mention plugin example code snippets: https://tiptap.dev/api/nodes/mention
-const suggestion = {
-  items: ({ query }: { query: string }) => {
-    return useVariablesStore
-      .getState()
-      .variables.filter((item) =>
-        item.toLowerCase().includes(query.toLowerCase()),
-      )
-      .slice(0, 5);
-  },
-
-  render: () => {
-    let component: undefined | ReactRenderer<any, any>;
-    let popup: Instance[] | undefined;
-
-    return {
-      onStart: (props: any) => {
-        component = new ReactRenderer(MentionList, {
-          props,
-          editor: props.editor,
-        });
-
-        if (!props.clientRect) {
-          return;
-        }
-
-        popup = tippy("body", {
-          getReferenceClientRect: props.clientRect,
-          appendTo: () => document.body,
-          content: component.element,
-          showOnCreate: true,
-          interactive: true,
-          trigger: "manual",
-          placement: "bottom-start",
-        });
-      },
-
-      onUpdate(props: any) {
-        component?.updateProps(props);
-
-        if (!props.clientRect) {
-          return;
-        }
-
-        popup?.[0]?.setProps({
-          getReferenceClientRect: props.clientRect,
-        });
-      },
-
-      onKeyDown(props: any) {
-        if (props.event.key === "Escape") {
-          popup?.[0]?.hide();
-          return true;
-        }
-
-        return component?.ref?.onKeyDown(props);
-      },
-
-      onExit() {
-        popup?.[0]?.destroy();
-        component?.destroy();
-      },
-    };
-  },
-};
-
-// Implemented based on the mention plugin example code snippets: https://tiptap.dev/api/nodes/mention
-const MentionList = forwardRef((props: MentionListProps, ref) => {
-  const variablesStore = useVariablesStore();
-  const [selectedIndex, setSelectedIndex] = useState(0);
-
-  const selectItem = (index: number) => {
-    const item = props.items[index];
-
-    if (item) {
-      props.command({ id: item });
-    }
-  };
-
-  const upHandler = () => {
-    setSelectedIndex(
-      (selectedIndex + props.items.length - 1) % props.items.length,
-    );
-  };
-
-  const downHandler = () => {
-    setSelectedIndex((selectedIndex + 1) % props.items.length);
-  };
-
-  const enterHandler = () => {
-    selectItem(selectedIndex);
-  };
-
-  useEffect(() => setSelectedIndex(0), [props.items]);
-
-  useImperativeHandle(ref, () => ({
-    onKeyDown: ({ event }: any) => {
-      if (event.key === "ArrowUp") {
-        upHandler();
-        return true;
-      }
-
-      if (event.key === "ArrowDown") {
-        downHandler();
-        return true;
-      }
-
-      if (event.key === "Enter") {
-        enterHandler();
-        return true;
-      }
-
-      return false;
-    },
-  }));
-
-  return (
-    <MentionItems>
-      {props.query.length > 0 && (
-        <MentionItemsButton
-          variant="text"
-          onClick={() => {
-            props.command({ id: props.query });
-            variablesStore.addVariable(props.query);
-          }}
-        >
-          + Add <span className={passportClassName}>@{props.query}</span>
-        </MentionItemsButton>
-      )}
-      {props.items.length > 0 ? (
-        props.items.map((item: any, index: number) => (
-          <MentionItemsButton
-            variant="text"
-            className={`mention-item ${
-              index === selectedIndex ? "mention-item-selected" : ""
-            }`}
-            key={index}
-            onClick={() => selectItem(index)}
-          >
-            <span className={passportClassName}>@{item}</span>
-          </MentionItemsButton>
-        ))
-      ) : (
-        <MentionItemsEmpty variant="body2">No results</MentionItemsEmpty>
-      )}
-    </MentionItems>
-  );
-});
 
 export default RichTextInput;
